Extract base class `Line` from `Statement`.
Make `Error` a subclass of `Line` and change the base class of `Blank`
from `Statement` to `Line`.

Change the type of the `CueSheet.lines` attribute from `Statement |
Error` to `Line`.  Drop the `errors` attribute and add an `errors`
property to `CueSheet`.

Simplify the return types of `parse_lines()` and `parse_line()`.