In the cuesheet parser, read a line of tokens at a time.
Since cuesheet commands always start with a command name and end with a
new line, lift the "line of tokens" reading code in `Parser.error()` up
into a new `Parser.next_line()` method, replacing the need to peek ahead
at tokens while parsing.

A line is either a well-formed command, an error or a blank line.  The
parser never needs to back-track.

Add a problematic cuesheet for future testing.